# Conflicts:
#	e2e/gui/src/test/java/com/epam/pipeline/autotests/LaunchParametersTest.java
#	e2e/gui/src/test/java/com/epam/pipeline/autotests/ao/SettingsPageAO.java
#	e2e/gui/src/test/java/com/epam/pipeline/autotests/utils/Json.java

+++ scripts/pipeline-launch/launch.sh
@@ -389,11 +389,26 @@ function configure_package_manager {
             if [ "$CP_OS" == "centos" ]; then
                   for _CP_REPO_RETRY_ITER in $(seq 1 $CP_REPO_RETRY_COUNT); do
                         curl -sk "${CP_REPO_BASE_URL}/cloud-pipeline.repo" > /etc/yum.repos.d/cloud-pipeline.repo && \
-                        yum --disablerepo=* --enablerepo=cloud-pipeline install yum-priorities -y -q > /dev/null 2>&1
-                        
+                        yum --disablerepo=* --enablerepo=cloud-pipeline install yum-priorities -y -q >> /var/log/yum.cp.log 2>&1
+
                         if [ $? -ne 0 ]; then
                               echo "[ERROR] (attempt: $_CP_REPO_RETRY_ITER) Failed to configure $CP_REPO_BASE_URL for the yum, removing the repo"
                               rm -f /etc/yum.repos.d/cloud-pipeline.repo
+                        else
+                              # If the CP repo was configured correctly - allow others fail
+                              if ! check_installed "yum-config-manager"; then
+                                    yum --disablerepo=* --enablerepo=cloud-pipeline install yum-utils -y -q >> /var/log/yum.cp.log 2>&1
+                              fi
+                              yum-config-manager --save --setopt=\*.skip_if_unavailable=true >> /var/log/yum.cp.log 2>&1
+                              # Disable "fastermirror" as it slows down the installtion and is not needed during the CP repo usage
+                              if [ -f "/etc/yum/pluginconf.d/fastestmirror.conf" ]; then
+                                    sed -i 's/enabled=1/enabled=0/g' /etc/yum/pluginconf.d/fastestmirror.conf
+                              fi
+                              # Use the "base" url for the other repos, as the mirrors may cause issues
+                              sed -i 's/^#baseurl=/baseurl=/g' /etc/yum.repos.d/*.repo
+                              sed -i 's/^metalink=/#metalink=/g' /etc/yum.repos.d/*.repo
+                              sed -i 's/^mirrorlist=/#mirrorlist=/g' /etc/yum.repos.d/*.repo
+                              break
                         fi
                   done
             elif [ "$CP_OS" == "debian" ] || [ "$CP_OS" == "ubuntu" ]; then
@@ -433,6 +448,11 @@ function get_install_command_by_current_distr {
             check_installed "yum" && _ltdl_lib_name="libtool-ltdl"
             _TOOLS_TO_INSTALL="$(sed "s/\( \|^\)ltdl\( \|$\)/ ${_ltdl_lib_name} /g" <<< "$_TOOLS_TO_INSTALL")"
       fi
+      if [[ "$_TOOLS_TO_INSTALL" == *"python"* ]] && \
+         [ "$CP_OS" == "centos" ] && \
+         [ "$CP_VER" == "8" ]; then
+            _TOOLS_TO_INSTALL="$(sed -e "s/python/python2/g" <<< "$_TOOLS_TO_INSTALL")"
+      fi
 
       local _TOOL_TO_CHECK=
       local _TOOLS_TO_INSTALL_VERIFIED=
@@ -447,9 +467,12 @@ function get_install_command_by_current_distr {
             _INSTALL_COMMAND_TEXT=
       else
             check_installed "apt-get" && { _INSTALL_COMMAND_TEXT="rm -rf /var/lib/apt/lists/; apt-get update -y -qq --allow-insecure-repositories; DEBIAN_FRONTEND=noninteractive apt-get -y -qq --allow-unauthenticated -o Dpkg::Options::=\"--force-confold\" install $_TOOLS_TO_INSTALL_VERIFIED";  };
+            check_installed "apk" && { _INSTALL_COMMAND_TEXT="apk update -q 1>/dev/null; apk -q add $_TOOLS_TO_INSTALL_VERIFIED";  };
             if check_installed "yum"; then
-                  check_installed "apk" && { _INSTALL_COMMAND_TEXT="apk update -q 1>/dev/null; apk -q add $_TOOLS_TO_INSTALL_VERIFIED";  };
-                  if [ "$CP_REPO_ENABLED" == "true" ] && [ -f /etc/yum.repos.d/cloud-pipeline.repo ]; then
+                  # Centos 8 throws "No available modular metadata for modular package" if all the other repos are disabled
+                  if [ "$CP_REPO_ENABLED" == "true" ] && \
+                     [ -f /etc/yum.repos.d/cloud-pipeline.repo ] && \
+                     [ "$CP_VER" != "8" ]; then
                         _INSTALL_COMMAND_TEXT="yum clean all -q && yum --disablerepo=* --enablerepo=cloud-pipeline -y -q install $_TOOLS_TO_INSTALL_VERIFIED"
                   else
                         _INSTALL_COMMAND_TEXT="yum clean all -q && yum -y -q install $_TOOLS_TO_INSTALL_VERIFIED"
@@ -980,6 +1003,11 @@ add_self_to_no_proxy
 # We need to make sure that the DIND and SYSTEMD are available if the Kubernetes is requested
 if check_cp_cap "CP_CAP_KUBE"; then
       export CP_CAP_DIND_CONTAINER="true"
+      export CP_CAP_SYSTEMD_CONTAINER="true"
+fi
+
+# We need to make sure that the Systemd is enabled if we use DCV
+if check_cp_cap "CP_CAP_DCV"; then
       export CP_CAP_SYSTEMD_CONTAINER="true"
 fi
 
@@ -1766,6 +1794,21 @@ if [ "$CP_CAP_DESKTOP_NM" == "true" ]; then
       nomachine_setup
 else
     echo "NoMachine support is not requested"
+fi
+
+######################################################
+
+######################################################
+# Setup Nice DCV
+######################################################
+
+echo "Setup NICE DCV environment"
+echo "-"
+
+if [ "$CP_CAP_DCV" == "true" ]; then
+      nice_dcv_setup
+else
+    echo "Nice DCV support is not requested"
 fi
 
 ######################################################
